jit3d: cast corrupted context back to tokens.dtype (bf16 autocast fix
torch.norm / division promote to float32 under autocast, but the original
tokens tensor is bf16 — direct index assignment raised a dtype mismatch.
Cast ctx_corrupted back to tokens.dtype on writeback.

+++ meteolibre_model/models/jit3d.py
@@ -162,7 +162,7 @@ class LatentContextCorruptor(nn.Module):
 
         # Write back only the context slice of masked samples
         out = tokens.clone()
-        out[mask, :n_ctx, :] = ctx_corrupted
+        out[mask, :n_ctx, :] = ctx_corrupted.to(tokens.dtype)
         return out
 
     def corrupt_embed(self, tokens: torch.Tensor, n_ctx: int) -> torch.Tensor:
